🚚 Move profile components to own folder
